- fixed compilation, removed warnings & updated codebase to newest JUCE version

+++ hi_backend/backend/debug_components/ScriptComponentPropertyPanels.cpp
@@ -696,7 +696,7 @@ void HiColourPropertyComponent::ColourComp::paint(Graphics& g)
 	auto r = getLocalBounds();
 	auto lr = r.removeFromLeft(80);
 
-	g.fillCheckerBoard(r, 10, 10, Colour(0xFF888888), Colour(0xFF444444));
+	g.fillCheckerBoard(r.toFloat(), 10, 10, Colour(0xFF888888), Colour(0xFF444444));
 
 	g.setColour(colour);
 	g.fillRect(r);
